Handle more ExtraConfig fields during resize
Support MMPowerOffVMExtraConfigKey and V1Alpha1ExtraConfigKey EC fields
during resize.

Rename the devops resize to just overwrite since we do set all of this
when we don't have a VM Class. This is another step towards being able
to replace the pre-power on reconfigure with resize.

Sanitize the VM Class ConfigSpec by removing the now deprecated
MMPowerOff EC field.
